Add HomePage component
Added HomePage Component; this component allows users to upload their game code as a file, which will then be sent to the backend for parsing. The backend returns a JSON object which contains the logic for the game.

NOTE: The backend currently returns static JSON as the backend functionality is not complete yet; this allows the game to continue functioning. Once backend functionality is complete, the server code will be changed.